requirements.txt
Flask==2.0.3
google-api-python-client==2.64.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0